Update remove-nondefaultshares.ps1
include ad shares

+++ scripts/remove-nondefaultshares.ps1
@@ -1,7 +1,7 @@
 Write-Output "`n---Removing every share besides ADMIN$, C$, and IPC$"
 $Shares = Get-SmbShare | Select-Object -ExpandProperty name
 foreach($Share in $Shares) {
-    if(($Share -ne "ADMIN$") -and ($Share -ne "C$") -and ($Share -ne "IPC$")) {
+    if(($Share -ne "ADMIN$") -and ($Share -ne "C$") -and ($Share -ne "IPC$") -and ($Share -ne "SYSVOL") -and ($Share -ne "NETLOGON")) {
         Write-Output "Removed share $Share"
         Remove-SmbShare -Name $Share
     }
